requirements.txt
celery
websockets
default-libmysqlclient-dev
build-essential
pkg-config
Django>=4.2.3,<5.0.0
django-rest-framework>=0.1.0
djangorestframework>=3.14.0
websocket-client>=1.6.1
numpy>=1.24.3
pandas>=2.0.2
requests>=2.31.0
python-binance>=1.0.19
cryptography>=41.0.1
pytz>=2023.3
Pillow>=10.0.0
whitenoise>=6.5.0
psycopg2-binary>=2.9.6
dj-database-url>=2.0.0
gunicorn>=21.2.0
python-telegram-bot>=20.4
twilio>=7.12.0
django-redis>=5.2.0
redis>=4.6.0
django-prometheus>=2.3.0
mysqlclient>=2.1.0
# Live chat dependencies
channels>=4.0.0
daphne>=4.0.0
channels-redis>=4.1.0
asgiref>=3.7.2
emoji>=2.10.0
markdown2>=2.4.10
bleach>=6.1.0
django-cors-headers>=4.3.1
# AI Agent dependencies
matplotlib>=3.7.1
openai>=1.2.0
pinecone-client>=6.0.0
langchain>=0.2.0
langchain-community>=0.3.20
langchain-pinecone>=0.2.3
langchain-openai>=0.1.0
faiss-cpu>=1.10.0
alpha_vantage